Allow tracing of debug adapter communication
Introduce the "debug.trace" preference.  When true, log the
communication between Theia and debug adapters to the output view.

I chose not to delete the output channels when the debug sessions close.
The reason being that if you are interested in what events led to the
end of a debug session, deleting the channel would make the interesting
output unavailable.  This seems reasonnable to me because:

- this feature is only meant to be used when debugging debug adapters
  and debug backends
- the output is only stored in the frontend and wiped on reload

Change-Id: I0f2a2557fa1ec0c61a21e7af3209c433bd9be755
Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>